fix(circular deps): refactor imports/exports to fix  circular dependency issues
