+++ packages/blockchain-wallet-v4-frontend/src/components/Flyout/RecurringBuy/RecurringBuyDetails.stories.tsx
@@ -2,8 +2,9 @@ import React from 'react'
 import { IntlProvider } from 'react-intl'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import { SBPaymentTypes } from '../../../../../blockchain-wallet-v4/src/network/api/simpleBuy/types'
-import { RecurringBuyPeriods } from '../../../data/types'
+import { SBPaymentTypes } from '@core/types'
+import { RecurringBuyPeriods } from 'data/types'
+
 import RecurringBuyDetails from './RecurringBuyDetails'
 
 export default {
