+++ packages/blockchain-wallet-v4-frontend/src/components/Flyout/model.tsx
@@ -2,11 +2,8 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 import moment, { Moment } from 'moment'
 
-import { PaymentMethod } from '@core/types'
-
-import { SBPaymentTypes } from '../../../../blockchain-wallet-v4/src/network/api/simpleBuy/types'
-import { RecurringBuyPeriods } from '../../data/components/recurringBuy/types'
-import { ActionEnum } from '../../data/types'
+import { SBPaymentTypes } from '@core/types'
+import { ActionEnum, RecurringBuyPeriods } from 'data/types'
 
 const RECURRING_BUY_PERIOD_FETCH = 'RECURRING_BUY_PERIOD_FETCH'
