+++ packages/blockchain-wallet-v4-frontend/src/data/coins/selectors/coins/custodial.tsx
@@ -5,7 +5,7 @@ import { ExtractSuccess } from '@core/remote/types'
 import { CoinType } from '@core/types'
 import { createDeepEqualSelector } from '@core/utils'
 import { generateTradingAccount } from 'data/coins/utils'
-import { SwapAccountType } from 'data/components/types'
+import { SwapAccountType } from 'data/types'
 
 import { getTradingBalance } from '..'
