+++ packages/blockchain-wallet-v4-frontend/src/data/modals/types.ts
@@ -1,6 +1,16 @@
-import { RecurringBuyOrigins } from 'data/components/types'
-
 import * as AT from './actionTypes'
+
+export enum RecurringBuyOrigins {
+  BUY_CONFIRMATION = 'BUY_CONFIRMATION',
+  COIN_PAGE = 'COIN_PAGE',
+  CURRENCY_PAGE = 'CURRENCY_PAGE',
+  DASHBOARD_PROMO = 'DASHBOARD_PROMO',
+  DETAILS_SCREEN = 'DETAILS_SCREEN',
+  RECURRING_BUYS_BANNER = 'RECURRING_BUYS_BANNER',
+  RECURRING_BUYS_FREQUENCY_SCREEN = 'RECURRING_BUYS_FREQUENCY_SCREEN',
+  SIMPLE_BUY_FREQUENCY_SCREEN = 'SIMPLE_BUY_FREQUENCY_SCREEN',
+  SIMPLE_BUY_ORDER_SUMMARY = 'SIMPLE_BUY_ORDER_SUMMARY'
+}
 
 export enum ModalName {
   ADD_BANK_YAPILY_MODAL = 'ADD_BANK_YAPILY_MODAL',
