--- packages/blockchain-wallet-v4-frontend/src/data/modules/types.ts
@@ -1,7 +1,0 @@
-import { ProfileActionTypes } from './profile/types'
-
-// All relative ActionTypes should go here
-// export type ModuleActionTypes = ProfileActionTypes | etc
-export type ModuleActionTypes = ProfileActionTypes
-
-export * from './profile/types'
