+++ packages/blockchain-wallet-v4-frontend/src/data/preferences/types.ts
@@ -1,5 +1,5 @@
 import { CoinType, SBOrderActionType, TimeRange } from '@core/types'
-import { SBFixType } from 'data/components/types'
+import { SBFixType } from 'data/types'
 
 import * as AT from './actionTypes'
