+++ packages/blockchain-wallet-v4-frontend/src/data/types.ts
@@ -1,8 +1,22 @@
 export * from './auth/types'
-export * from './components/types'
+export * from './components/brokerage/types'
+export * from './components/fundRecovery/types'
+export * from './components/identityVerification/types'
+export * from './components/interest/types'
+export * from './components/interestUploadDocument/types'
+export * from './components/recoveryPhrase/types'
+export * from './components/recurringBuy/types'
+export * from './components/request/types'
+export * from './components/sendBch/types'
+export * from './components/sendBtc/types'
+export * from './components/settings/types'
+export * from './components/simpleBuy/types'
+export * from './components/swap/types'
+export * from './components/withdraw/types'
 export * from './custodial/types'
 export * from './goals/types'
 export * from './modals/types'
-export * from './modules/types'
+export * from './modules/profile/types'
+export * from './modules/transferEth/types'
 export * from './prices/types'
 export * from './session/types'
