+++ packages/blockchain-wallet-v4-frontend/src/modals/InterestUploadDocuments/index.tsx
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react'
 import { connect, ConnectedProps } from 'react-redux'
 import { bindActionCreators, compose, Dispatch } from 'redux'
 
-import { FileUploadItem, RemoteDataType } from 'blockchain-wallet-v4/src/types'
+import { FileUploadItem, RemoteDataType } from '@core/types'
 import Flyout, { duration, FlyoutChild } from 'components/Flyout'
 import { actions } from 'data'
 import { RootState } from 'data/rootReducer'
